feat: add authorization on mobile and web and create auth token

+++ server/src/routes/memories.ts
@@ -14,9 +14,15 @@ const bodySchema = z.object({
 
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  
-  app.get('/memories', async (r) => {
+  app.addHook('preHandler', async (req) => {
+    await req.jwtVerify()
+  })
+
+  app.get('/memories', async (req) => {
     const memories = await prisma.memory.findMany({
+      where:{
+        userId: req.user.sub,
+      },
       orderBy: {
         createdAt: 'asc'
       }
@@ -31,7 +37,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (req) => {
+  app.get('/memories/:id', async (req, reply) => {
     const { id } = paramsSchema.parse(req.params)
 
     const memory = await prisma.memory.findUniqueOrThrow({
@@ -39,6 +45,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       }
     })
+
+    if(!memory.isPublic && memory.userId !== req.user.sub){
+      return reply.status(401).send()
+    }
 
     return memory
   })
@@ -51,18 +61,28 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '5be7f607-54f2-4a4e-85b0-f992615efef9'
+        userId: req.user.sub
       }
     })
 
     return memory
   })
 
-  app.put('/memories/:id', async (req) =>{
+  app.put('/memories/:id', async (req, reply) =>{
     const { id } = paramsSchema.parse(req.params)
     const {content, coverUrl, isPublic} = bodySchema.parse(req.body)
+
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+
+    if(memory.userId !== req.user.sub){
+      return reply.status(401).send()
+    }
     
-    const memory = await prisma.memory.update({
+      memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -76,8 +96,18 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.delete('/memories/:id', async (req) =>{
+  app.delete('/memories/:id', async (req, reply) =>{
     const { id } = paramsSchema.parse(req.params)
+    
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+
+    if(memory.userId !== req.user.sub){
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
